# openapi.json
{"openapi":"3.1.0","info":{"title":"Real Analytics Backend","version":"0.1.0"},"paths":{"/api/v1/datasets":{"get":{"tags":["Datasets"],"summary":"List Datasets","operationId":"list_datasets_api_v1_datasets_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},{"name":"source","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}},{"name":"updated_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated From"}},{"name":"updated_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated To"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"description":"Items per page","default":25,"title":"Page Size"},"description":"Items per page"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Datasets"],"summary":"Create Dataset","operationId":"create_dataset_api_v1_datasets_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}":{"get":{"tags":["Datasets"],"summary":"Get Dataset","operationId":"get_dataset_api_v1_datasets__dataset_id__get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Datasets"],"summary":"Update Dataset","operationId":"update_dataset_api_v1_datasets__dataset_id__put","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Datasets"],"summary":"Delete Dataset","operationId":"delete_dataset_api_v1_datasets__dataset_id__delete","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/{dataset_id}/files":{"post":{"tags":["Datasets"],"summary":"Register File","operationId":"register_file_api_v1_datasets__dataset_id__files_post","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetFileCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Datasets"],"summary":"List Files","operationId":"list_files_api_v1_datasets__dataset_id__files_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/run":{"post":{"tags":["Analysis"],"summary":"Run Analysis","operationId":"run_analysis_api_v1_analysis_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisRunRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/jobs/{job_id}":{"get":{"tags":["Analysis"],"summary":"Job Status","operationId":"job_status_api_v1_analysis_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/jobs/{job_id}/results":{"get":{"tags":["Analysis"],"summary":"Job Results","operationId":"job_results_api_v1_analysis_jobs__job_id__results_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/statistics/summary":{"post":{"tags":["Statistics"],"summary":"Summary","operationId":"summary_api_v1_statistics_summary_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatisticsSummaryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/features/queue":{"post":{"tags":["Features"],"summary":"Queue Feature Operation","operationId":"queue_feature_operation_api_v1_features_queue_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureQueueRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/features/{operation_id}":{"get":{"tags":["Features"],"summary":"Get Operation","operationId":"get_operation_api_v1_features__operation_id__get","parameters":[{"name":"operation_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Operation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/experiments":{"get":{"tags":["Experiments"],"summary":"List Experiments","operationId":"list_experiments_api_v1_experiments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Experiments"],"summary":"Create Experiment","operationId":"create_experiment_api_v1_experiments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperimentCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/experiments/{experiment_id}":{"get":{"tags":["Experiments"],"summary":"Get Experiment","operationId":"get_experiment_api_v1_experiments__experiment_id__get","parameters":[{"name":"experiment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Experiment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Experiments"],"summary":"Update Experiment","operationId":"update_experiment_api_v1_experiments__experiment_id__patch","parameters":[{"name":"experiment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Experiment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperimentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Experiments"],"summary":"Delete Experiment","operationId":"delete_experiment_api_v1_experiments__experiment_id__delete","parameters":[{"name":"experiment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Experiment Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/experiments/{experiment_id}/runs":{"get":{"tags":["Experiments"],"summary":"Experiment Runs","operationId":"experiment_runs_api_v1_experiments__experiment_id__runs_get","parameters":[{"name":"experiment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Experiment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/dashboard":{"get":{"tags":["Analytics"],"summary":"Dashboard","operationId":"dashboard_api_v1_analytics_dashboard_get","parameters":[{"name":"dataset_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"}},{"name":"experiment_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Experiment Id"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/{job_id}/logs":{"get":{"tags":["Jobs"],"summary":"Job Logs","operationId":"job_logs_api_v1_jobs__job_id__logs_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/metrics":{"get":{"summary":"Metrics","description":"Endpoint that serves Prometheus metrics.","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnalysisRunRequest":{"properties":{"dataset_id":{"type":"string","format":"uuid","title":"Dataset Id"},"analysis_types":{"items":{"$ref":"#/components/schemas/AnalysisType"},"type":"array","title":"Analysis Types"},"options":{"additionalProperties":true,"type":"object","title":"Options"}},"type":"object","required":["dataset_id","analysis_types"],"title":"AnalysisRunRequest"},"AnalysisType":{"type":"string","enum":["descriptive","quality","correlation","timeseries","distribution"],"title":"AnalysisType"},"DatasetCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description","default":""},"schema":{"additionalProperties":true,"type":"object","title":"Schema"},"row_count":{"type":"integer","title":"Row Count","default":0},"source":{"$ref":"#/components/schemas/DatasetSource","default":"internal"},"files":{"items":{"$ref":"#/components/schemas/DatasetFileCreate"},"type":"array","title":"Files"}},"type":"object","required":["name"],"title":"DatasetCreate"},"DatasetFileCreate":{"properties":{"path":{"type":"string","title":"Path"},"checksum":{"type":"string","title":"Checksum"},"file_type":{"type":"string","title":"File Type","default":"csv"}},"type":"object","required":["path","checksum"],"title":"DatasetFileCreate"},"DatasetSource":{"type":"string","enum":["internal","external","synthetic"],"title":"DatasetSource"},"DatasetUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Schema"},"row_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Row Count"},"source":{"anyOf":[{"$ref":"#/components/schemas/DatasetSource"},{"type":"null"}]}},"type":"object","title":"DatasetUpdate"},"ExperimentCreate":{"properties":{"dataset_id":{"type":"string","format":"uuid","title":"Dataset Id"},"name":{"type":"string","title":"Name"},"algorithm":{"type":"string","title":"Algorithm"},"hyperparameters":{"additionalProperties":true,"type":"object","title":"Hyperparameters"}},"type":"object","required":["dataset_id","name","algorithm"],"title":"ExperimentCreate"},"ExperimentStatus":{"type":"string","enum":["draft","running","completed","failed","cancelled"],"title":"ExperimentStatus"},"ExperimentUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"algorithm":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Algorithm"},"hyperparameters":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Hyperparameters"},"status":{"anyOf":[{"$ref":"#/components/schemas/ExperimentStatus"},{"type":"null"}]}},"type":"object","title":"ExperimentUpdate"},"FeatureOperationType":{"type":"string","enum":["impute","encode","scale","normalize"],"title":"FeatureOperationType"},"FeatureQueueRequest":{"properties":{"dataset_id":{"type":"string","format":"uuid","title":"Dataset Id"},"operations":{"items":{"$ref":"#/components/schemas/FeatureOperationType"},"type":"array","title":"Operations"},"config":{"additionalProperties":true,"type":"object","title":"Config"}},"type":"object","required":["dataset_id","operations"],"title":"FeatureQueueRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"StatisticsSummaryRequest":{"properties":{"dataset_id":{"type":"string","format":"uuid","title":"Dataset Id"},"columns":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Columns"},"options":{"additionalProperties":true,"type":"object","title":"Options"}},"type":"object","required":["dataset_id"],"title":"StatisticsSummaryRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}